refactor tsp-openapi3 test-host
